chore: chore: clean up positions
